Warning Removal and Conversion Bug Squash
Now compiles with "-Wall -Wextra -Wpedantic" meaning that every "common" warning from GCC has been addressed. The commit also fixes a bug in the BMP-MHS conversion code. However, what that bug was and how it was fixed are questions I cannot answer as the bug seemed to magically be fixed with the edits I made.