Convert Update
This commit adds the ability to convert from an MHS file to either a skeleton and palette or a set of colored images.
Also included in this commit is an updated README for the new changes and the ability for Minif text to be scaled larger on the screen. This second change is visible in "sdl-test E" and has had controls added to the "display" mode to resize the info text.