Two Major Fixes
This commit addresses two massives bugs in the colormap copying routine. The first was a double-free that would arise later in execution due to freeing `colorLUT` when leaving the scope despite assigning the pointer to the header. The second was that the color data wouldn't actually be read from the file, which would lead to garbage data being read for the color values.